misc: Use other size modifiers to fix compiler warnings
control_if.c:521:2: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 3 has type ‘uint64_t’ [-Wformat]
osmo_bsc_bssap.c:473:3: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 7 has type ‘unsigned int’ [-Wformat]
mgcp_main.c:162:4: warning: format ‘%lu’ expects argument of type ‘long unsigned int’, but argument 4 has type ‘unsigned int’ [-Wformatt]

+++ openbsc/src/osmo-bsc_mgcp/mgcp_main.c
@@ -158,8 +158,8 @@ static int read_call_agent(struct osmo_fd *fd, unsigned int what)
 		perror("Gateway failed to read");
 		return -1;
 	} else if (slen > sizeof(addr)) {
-		fprintf(stderr, "Gateway received message from outerspace: %lu %lu\n",
-			(unsigned long int)slen, sizeof(addr));
+		fprintf(stderr, "Gateway received message from outerspace: %zu %zu\n",
+			slen, sizeof(addr));
 		return -1;
 	}
